move TextInput to components/core
