+++ src/components/legacy/TextInput/TextInput.js
@@ -159,7 +159,7 @@ const InputLabel = styled(MediumText)`
 /**
  * @deprecated This compontent is considered legacy and should not be used in new code
  *
- * Use: components/inputs/TextInput instead
+ * Use: components/core/TextInput instead
  */
 class TextInput extends React.Component<Props, State> {
   multilineInputField: Input;
